requirements.txt
alabaster==0.7.10
amqp==2.2.1
aniso8601==1.2.1
appnope==0.1.0
Babel==2.4.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.3
bleach==2.0.0
boto3==1.4.4
botocore==1.5.91
celery==4.1.0
certifi==2017.4.17
click==6.7
colorama==0.3.9
configparser==3.5.0
decorator==4.0.11
Django==1.11.12
docopt==0.6.2
docutils==0.13.1
entrypoints==0.2.2
enum34==1.1.6
Flask==0.12.2
Flask-BasicAuth==0.2.0
Flask-RESTful==0.3.6
Flask-Session==0.3.1
functools32==3.2.3.post2
futures==3.1.1
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
html5lib==0.999999999
image==1.5.20
imagesize==0.7.1
inifile==0.3
ipykernel==4.6.1
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.6.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
kaptan==0.5.8
kombu==4.1.0
libtmux==0.7.3
MarkupSafe==1.0
minio==2.2.4
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.3.0
neo==0.5.1
networkx==1.11
notebook==5.0.0
numpy==1.11.3
pandas==0.19.2
pandocfilters==1.4.1
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.14
protobuf==3.2.0
ptyprocess==0.5.1
pydicom==1.0.0a1
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.0
quantities==0+unknown
redis==2.10.5
requests==2.13.0
requests-futures==0.9.7
retrying==1.3.3
s3transfer==0.1.10
scandir==1.5
scikit-learn==0.18.2
scipy==0.19.1
simplegeneric==0.8.1
simplejson==3.11.1
singledispatch==3.4.0.3
six==1.10.0
sklearn==0.0
snowballstemmer==1.2.1
spark==0.2.1
Sphinx==1.6.2
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
terminado==0.6
testpath==0.3.1
tmuxp==1.3.1
tornado==4.5.1
traitlets==4.3.2
typing==3.6.1
urllib3==1.21.1
validator.py==1.2.5
vine==1.1.4
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==2.0.0
yasm==0.0
